requirements.txt
discord.py>=2.6
python-dotenv>=1.0
asyncmy>=0.2.9
tabulate==0.9.0